refactor: Move common functions and schemas to helpers
